Last prompts update to match spec
